fix(solver): renamed agent folder UseProblemSolvingMethodModule
