+++ problem-solver/py/modules/UseProblemSolvingMethodModule/dataClasses/AnnStruct.py
