+++ problem-solver/py/modules/UseProblemSolvingMethodModule/dataClasses/InterpretationParameters.py
@@ -3,14 +3,14 @@ from typing import List
 
 from sc_client.models import ScAddr
 
-from inputOutput.InputOutputStruct import InputOutput
-from modules.fnnUseProblemSolvingMethodModule.dataClasses.AnnStruct import AnnStruct
+from .AnnStruct import AnnStruct
+from .inputOutput.ImageStruct import Image
 
 
 @dataclass
 class InterpretationParameters:
     ann_struct: AnnStruct
-    input_struct: InputOutput               # inputs for ann
+    input_struct: Image               # inputs for ann
     problem_types: List[ScAddr] = None      # todo or maybe unused
